delete deprecated WP blog handling
