+++ frontend/package.json
@@ -14,11 +14,11 @@
 	},
 	"dependencies": {
 		"@emotion/styled": "^11.14.0",
-		"@fontsource-variable/dm-sans": "^5.2.5",
-		"@fontsource-variable/inter": "^5.2.5",
-		"@fontsource/roboto": "^5.2.5",
-		"@fontsource/roboto-condensed": "^5.2.5",
-		"@fontsource/taviraj": "^5.2.5",
+		"@fontsource-variable/dm-sans": "^5.2.8",
+		"@fontsource-variable/inter": "^5.2.8",
+		"@fontsource/roboto": "^5.2.10",
+		"@fontsource/roboto-condensed": "^5.2.8",
+		"@fontsource/taviraj": "^5.2.8",
 		"@mui/icons-material": "^7.3.3",
 		"@mui/material": "^7.3.3",
 		"@tanstack/query-sync-storage-persister": "^5.90.2",
