+++ frontend/playwright-tests/externalUrls.spec.ts
@@ -49,15 +49,6 @@ const knownFlakyUrls = [
 
 test.describe.configure({ mode: 'parallel' })
 
-// TEST Headless Wordpress Endpoint
-test(`Fetch First 100 Blog Posts`, async ({ page }) => {
-  const url =
-    'https://hetblog.dreamhosters.com/wp-json/wp/v2/posts?_embed&per_page=100'
-  const response = await page.goto(url, { waitUntil: 'domcontentloaded' })
-  if (response?.status() !== 200)
-    console.error('\n🙀', url, response?.status(), '\n')
-})
-
 for (const url of Object.values(urlMap)) {
   if (!url || knownFlakyUrls.includes(url)) continue
   test(`${url}`, async ({ page }) => {
