+++ frontend/src/index.css
@@ -154,245 +154,6 @@ button[data-action="back"]:focus {
 	outline: 1px solid var(--color-white);
 }
 
-/* STYLING FOR FETCHED WORDPRESS HTML FOR NEWS AND STORIES */
-.fetched-wordpress-html {
-	> div {
-		display: flex;
-		flex-direction: column;
-		align-items: center;
-		margin: 0 auto;
-		padding: 2rem 8rem;
-		width: 100%;
-		max-width: 1440px !important;
-		justify-content: center !important;
-
-		@media (width >= theme(--breakpoint-xs)) {
-			padding: 0.5rem 1rem;
-		}
-
-		@media (width >= theme(--breakpoint-sm)) {
-			padding: 2rem 3rem;
-		}
-		@media (width >= theme(--breakpoint-md)) {
-			padding: 2rem 4rem;
-		}
-	}
-
-	& figure,
-	div > figure > img {
-		margin: 0 auto !important;
-		width: 100%;
-		max-width: 80vw;
-		height: auto;
-		display: flex;
-		flex-direction: column;
-		align-items: center;
-		justify-content: start;
-	}
-
-	& figure > a > img {
-		display: block;
-		min-width: 70vw;
-		max-width: 88vw;
-		margin: 1rem auto;
-		height: auto;
-	}
-
-	& figure > figcaption {
-		margin: 0.5rem 2rem;
-		padding: 0 2rem;
-		max-width: 88vw;
-		text-align: center !important;
-		font-size: 0.8rem !important;
-		font-weight: 300;
-		text-decoration: none !important;
-		font-style: italic;
-
-		@media (width >= theme(--breakpoint-xs)) {
-			padding: 0;
-			margin: 0 auto;
-		}
-	}
-
-	& figcaption {
-		margin-top: 0.5rem;
-		margin-bottom: 4rem;
-		text-align: center !important;
-		font-size: 0.8rem !important;
-		font-weight: 300;
-		text-decoration: none !important;
-		font-style: italic;
-		max-width: 600px;
-	}
-
-	& p {
-		width: 100%;
-		max-width: 95vw;
-		font-size: 1rem;
-	}
-
-	& p > a,
-	p > strong > a,
-	p > em > a,
-	div > span > a {
-		text-decoration: none;
-		font-weight: 700;
-	}
-
-	& p > a:hover {
-		text-decoration: underline;
-	}
-
-	& p > strong {
-		margin: 1rem 0;
-		font-weight: 600;
-		font-size: var(--text-title) !important;
-		font-family: var(--font-sans-title) !important;
-	}
-
-	& ul,
-	& ol,
-	& h1,
-	& h2,
-	& h3,
-	& h4,
-	& h5,
-	& h6 {
-		@media (width >= theme(--breakpoint-sm)) {
-			width: 100%;
-		}
-		@media (width >= theme(--breakpoint-md)) {
-			width: 95%;
-		}
-	}
-
-	& ul {
-		list-style: disc;
-		margin-left: 1rem;
-		padding-left: 1rem;
-		font-size: 1rem;
-	}
-
-	& hr {
-		margin-block: 3rem;
-		width: 50%;
-	}
-
-	& blockquote {
-		width: 100%;
-		max-width: 88vw;
-		padding: 2rem;
-	}
-
-	& blockquote > p {
-		font-style: italic;
-		display: flex;
-		flex-direction: column;
-		align-items: center;
-		margin: 0 auto;
-		justify-content: center !important;
-	}
-
-	& h1,
-	h1 > a {
-		font-family: "DM Sans Variable", sans-serif !important;
-		font-size: var(--text-header) !important;
-		line-height: 110%;
-		color: var(--color-alt-green);
-		font-weight: 700;
-		margin-top: 3rem;
-		margin-bottom: 0rem;
-		text-align: left;
-		text-decoration: none;
-	}
-
-	& h2,
-	h2 > a,
-	h2 a {
-		font-family: "DM Sans Variable", sans-serif !important;
-		font-size: var(--text-header) !important;
-		line-height: 110%;
-		color: var(--color-alt-green);
-		font-weight: 700;
-		margin-top: 3rem;
-		margin-bottom: 0rem;
-		text-align: left;
-		text-decoration: none;
-	}
-
-	& h3,
-	h3 > a {
-		font-family: "DM Sans Variable", sans-serif !important;
-		font-size: var(--text-header) !important;
-		line-height: 110%;
-		color: var(--color-alt-green);
-		font-weight: 300;
-		margin-top: 3rem;
-		margin-bottom: 0rem;
-		text-align: left;
-	}
-
-	& h4,
-	h4 > a {
-		font-family: "DM Sans Variable", sans-serif !important;
-		font-size: var(--text-title) !important;
-		line-height: 110%;
-		color: var(--color-alt-green);
-		font-weight: 500;
-		margin-top: 1rem;
-		margin-bottom: 0rem;
-		text-align: left;
-	}
-
-	& h5,
-	h5 > a {
-		font-family: "DM Sans Variable", sans-serif !important;
-		font-size: var(--text-title) !important;
-		line-height: 110%;
-		color: var(--color-alt-green);
-		font-weight: 300;
-		margin-top: 1rem;
-		margin-bottom: 0rem;
-		text-align: left;
-	}
-
-	& h6,
-	h6 > a {
-		font-family: "DM Sans Variable", sans-serif !important;
-		font-size: var(--text-text) !important;
-		line-height: 110%;
-		color: var(--color-alt-green);
-		font-weight: 300;
-		margin-top: 1rem;
-		margin-bottom: 0rem;
-		text-align: left;
-	}
-
-	/* hide featured image since already in header; hide HRs; show youtube embed <figure> */
-	.wp-block-post-featured-image {
-		display: none !important;
-	}
-
-	/* set youtube video container to proper aspect ratio */
-	.wp-block-embed-youtube .wp-block-embed__wrapper {
-		margin-block: 3rem;
-		position: relative;
-		width: 100%;
-		height: 0;
-		padding-bottom: 56.25%;
-	}
-
-	/* fill iframe to above container */
-	.wp-block-embed-youtube .wp-block-embed__wrapper > iframe {
-		position: absolute;
-		width: 100%;
-		height: 100%;
-		left: 0;
-		top: 0;
-		border-radius: 20px;
-	}
-}
-
 /* Base styles without animation */
 .fade-in-up-blur {
 	visibility: visible; /* Default visible state for reduced motion */
