+++ frontend/src/pages/ShareYourStory/CardLeftNumber.tsx
