+++ frontend/src/pages/ShareYourStory/ShareYourStoryContent.tsx
