+++ frontend/src/pages/ShareYourStory/ShareYourStoryPage.tsx
