+++ frontend/src/utils/externalUrls.ts
@@ -1,3 +1,5 @@
+import { SATCHER_NEWS_PAGE } from './blogUtils'
+
 export const urlMap: Record<string, string> = {
   acs5: 'https://www.census.gov/data/developers/data-sets/acs-5year.html',
   ahr: 'https://www.americashealthrankings.org/explore/measures',
@@ -20,6 +22,8 @@ export const urlMap: Record<string, string> = {
   hetLinkedIn: 'https://www.linkedin.com/company/healthequitytracker/',
   hetTikTok: 'https://www.tiktok.com/@healthequitytracker',
   hetYouTubeShorts: 'https://www.youtube.com/@HealthEquityTracker/shorts',
+  hivCriminalization:
+    SATCHER_NEWS_PAGE + '/hiv-criminalization-in-georgia-a-call-for-reform',
   lifeline: 'https://988lifeline.org',
   msm: 'https://www.msm.edu/',
   newsletterSignup:
